Made BitArray work as a bit array
It seems you might have gotten confused by the problem.

A bit is not a single number in the output string (e.g. the "5" in "5102"), it is instead a single digit in the binary number system (see: http://en.wikipedia.org/wiki/Bit).

As such, the requested project here seems to be to create an array of bits which will return the binary value at a certain index from a 64-bit number.

I have edited the code to reflect this new scope.